refactor: bibleVersionSelect feature로 분리
