--- src/features/drilling/components/bibleVersionSelect/index.test.tsx
@@ -1,3 +1,0 @@
-import { test } from 'vitest';
-
-test('BibleVersionSelect listbox renders bible version options', () => {});
